+++ lib/event_bloc_tester.dart
@@ -1,3 +1,4 @@
+/// The general tester for writing unit tests
 library event_bloc_tester;
 
 export '/src/tester.dart';