+++ lib/event_bloc_widget_tester.dart
@@ -1,4 +1,5 @@
+/// The general tester for writing unit tests
 library event_bloc_widget_tester;
 
-export 'event_bloc_tester.dart';
 export '/src/widget_tester.dart';
+export 'event_bloc_tester.dart';