+++ lib/src/widget_tester.dart
@@ -1,46 +1,56 @@
 import 'dart:async';
-import 'package:flutter_test/flutter_test.dart';
 
-import '../event_bloc_tester.dart';
+import 'package:event_bloc_tester/event_bloc_tester.dart';
+import 'package:flutter_test/flutter_test.dart';
 
 /// This helps you easily write widget tests that have a serializable output.
 ///
-/// Based on the [ListTesterMode] you use, this will either generate the expected output or
-/// test against the expected output generated with a previous [generateOuput] call.
+/// Based on the [ListTesterMode] you use, this will either generate the
+/// expected output or test against the expected output generated with a
+/// previous [ListTesterMode.generateOutput] call.
 ///
 /// Please see test/example_widget_test.dart for some example usage of this class.
 ///
 /// If you need a version for unit tests, please see [SerializableListTester]
 class SerializableListWidgetTester<T> with SerializableListTesterMixin<T> {
+  /// [testGroupName] and [mainTestName] are passed as the test groups this
+  /// tester is a part of.
+  ///
+  /// [mode] determines the behaviour of [runTests]. If it's in
+  /// [ListTesterMode.generateOutput], an output file will be generated based on
+  /// the output. If it's in [ListTesterMode.testOutput] the output will be
+  /// tested against the output file generated with a previous run of
+  /// [ListTesterMode.generateOutput]
+  ///
+  /// [testMap] is a map with key being the test name of the test, and the value
+  /// being a supplier for an initial testValue.
+  ///
+  /// [testFunction] is the common test that will be run on everything in
+  /// [testMap]. Use the provided [SerializableTester] to add your test values.
+  SerializableListWidgetTester({
+    required this.testGroupName,
+    required this.mainTestName,
+    required this.testFunction,
+    required this.testMap,
+    this.mode = ListTesterMode.testOutput,
+  });
   @override
   final String testGroupName;
   @override
   final String mainTestName;
   @override
   final ListTesterMode mode;
-  final FutureOr<void> Function(
-          T testInput, SerializableTester tester, WidgetTester widgetTester)
-      testFunction;
-  final Map<String, T Function()> testMap;
 
-  /// [testGroupName] and [mainTestName] are passed as the test groups this tester is a part of.
-  ///
-  /// [mode] determines the behaviour of [runTests]. If it's in [ListTesterMode.generateOutput],
-  /// an output file will be generated based on the output. If it's in [ListTesterMode.testOutput]
-  /// the output will be tested against the output file generated with a previous run of [ListTesterMode.generateOutput]
-  ///
-  /// [testMap] is a map with key being the test name of the test, and the value being a supplier for an
-  /// initial testValue.
-  ///
-  /// [testFunction] is the common test that will be run on everything in [testMap]. Use the provided
-  /// [SerializableTester] to add your test values.
-  SerializableListWidgetTester({
-    required this.testGroupName,
-    required this.mainTestName,
-    this.mode = ListTesterMode.testOutput,
-    required this.testFunction,
-    required this.testMap,
-  });
+  /// The function that will be ran for each test case in [testMap]
+  final FutureOr<void> Function(
+    T testInput,
+    SerializableTester tester,
+    WidgetTester widgetTester,
+  ) testFunction;
+
+  /// Holds the test cases with the test name being the key and the supplier
+  /// function for the test case being the value.
+  final Map<String, T Function()> testMap;
 
   @override
   Future<void> runTests() async {
@@ -48,15 +58,18 @@ class SerializableListWidgetTester<T> with SerializableListTesterMixin<T> {
 
     for (final testName in testMap.keys) {
       final completer = Completer<SerializableTester>();
-      testWidgets(testName,
-          (widgetTester) => runTest(path, testName, widgetTester, completer));
+      testWidgets(
+        testName,
+        (widgetTester) => runTest(path, testName, widgetTester, completer),
+      );
 
-      completer.future.then((value) => value.finish(path, testName));
+      unawaited(completer.future.then((value) => value.finish(path, testName)));
     }
 
     if (mode == ListTesterMode.generateOutput || mode == ListTesterMode.auto) {
-      // This is super janky, but we need it so that the SerializableTester will output to local.
-      testWidgets("Delay to let file output happen...", (widgetTester) async {
+      // This is super janky, but we need it so that the SerializableTester will
+      // output to local.
+      testWidgets('Delay to let file output happen...', (widgetTester) async {
         final currentTime = DateTime.now();
         while (DateTime.now().millisecondsSinceEpoch -
                 currentTime.millisecondsSinceEpoch <
@@ -67,6 +80,13 @@ class SerializableListWidgetTester<T> with SerializableListTesterMixin<T> {
     }
   }
 
+  /// Runs a single test given the [path] and [testName] to determine the
+  /// output location and [mode] to determine what to do with the output.
+  ///
+  /// [widgetTester] should be provided by [testWidgets]
+  ///
+  /// [completer] will be completed with the [SerializableTester] generated
+  /// after [testFunction] has been run.
   Future<void> runTest(
     String path,
     String testName,